requirements.txt
asgiref==3.7.2
Brotli==1.1.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dj-database-url==2.1.0
Django==5.0.2
django-cors-headers==4.3.1
django-rest-authtoken==2.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
fonttools==4.48.1
gunicorn==21.2.0
h11==0.14.0
inflection==0.5.1
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.2
mpld3==0.5.10
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pillow==10.2.0
plotly==5.18.0
psycopg2==2.9.9
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
seaborn==0.13.2
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.3
typing_extensions==4.10.0
tzdata==2023.4
uritemplate==4.1.1
uvicorn==0.29.0
waitress==3.0.0
whitenoise==6.6.0
